chore: port remaining types
