+++ packages/channel-client/package.json
@@ -38,6 +38,7 @@
     "typescript": "^3.7.3"
   },
   "dependencies": {
+    "@statechannels/client-api-schema": "0.0.1",
     "@statechannels/channel-provider": "^0.0.3",
     "ethers": "^4.0.40"
   },
