+++ packages/channel-client/src/types.ts
@@ -1,3 +1,11 @@
+import {
+  JsonRpcErrorResponse,
+  JsonRpcError,
+  JsonRpcRequest,
+  JsonRpcResponse,
+  JsonRpcNotification
+} from '@statechannels/channel-provider';
+
 export type ChannelStatus = 'proposed' | 'opening' | 'funding' | 'running' | 'closing' | 'closed';
 
 export interface Participant {
@@ -41,10 +49,6 @@ export interface ChannelResult {
 
 export type UnsubscribeFunction = () => void;
 
-export interface PushMessageResult {
-  success: boolean;
-}
-
 // The message Payload is designed to be opaque to the app. However, it's useful
 // to be able to specify the Payload type for the FakeChannelClient, as we'll be
 // manipulating it within the client.
@@ -78,3 +82,139 @@ export interface EventsWithArgs {
 }
 
 type UnsubscribeFunction = () => void;
+
+interface CreateChannelParameters {
+  participants: Participant[];
+  allocations: Allocation[];
+  appDefinition: string;
+  appData: string;
+}
+
+interface UpdateChannelParameters {
+  channelId: string;
+  participants: Participant[];
+  allocations: Allocation[];
+  appData: string;
+}
+
+export type GetAddressRequest = JsonRpcRequest<'GetAddress', {}>; // todo: what are params
+
+export type CreateChannelRequest = JsonRpcRequest<'CreateChannel', CreateChannelParameters>;
+
+export type CreateChannelResponse = JsonRpcResponse<ChannelResult>;
+
+export interface JoinChannelParameters {
+  channelId: string;
+}
+
+export type JoinChannelRequest = JsonRpcRequest<'JoinChannel', JoinChannelParameters>;
+
+export type JoinChannelResponse = JsonRpcResponse<ChannelResult>;
+
+export type UpdateChannelRequest = JsonRpcRequest<'UpdateChannel', UpdateChannelParameters>;
+
+export type PushMessageRequest = JsonRpcRequest<'PushMessage', Message>;
+
+export interface PushMessageResult {
+  success: boolean;
+}
+
+export type PushMessageResponse = JsonRpcResponse<PushMessageResult>;
+
+export interface CloseChannelParameters {
+  channelId: string;
+}
+export type CloseChannelRequest = JsonRpcRequest<'CloseChannel', CloseChannelParameters>;
+export type CloseChannelResponse = JsonRpcResponse<ChannelResult>;
+
+export type ChannelProposedNotification = JsonRpcNotification<'ChannelProposed', ChannelResult>;
+export type ChannelUpdatedNotification = JsonRpcNotification<'ChannelUpdated', ChannelResult>;
+export type ChannelClosingNotification = JsonRpcNotification<'ChannelClosed', ChannelResult>;
+
+export type MessageQueuedNotification = JsonRpcNotification<'MessageQueued', Message>;
+
+export type Notification =
+  | ChannelProposedNotification
+  | ChannelUpdatedNotification
+  | ChannelClosingNotification
+  | MessageQueuedNotification;
+
+export type NotificationName = Notification['method'];
+
+export type Request =
+  | GetAddressRequest
+  | CreateChannelRequest
+  | JoinChannelRequest
+  | UpdateChannelRequest
+  | PushMessageRequest
+  | CloseChannelRequest;
+
+export enum ErrorCodes {
+  SIGNING_ADDRESS_NOT_FOUND = 1000,
+  INVALID_APP_DEFINITION = 1001,
+  INVALID_APP_DATA = 1002,
+  UNSUPPORTED_TOKEN = 1003,
+  CHANNEL_NOT_FOUND = 1004
+}
+
+export class ChannelClientError implements JsonRpcErrorResponse {
+  jsonrpc: '2.0' = '2.0';
+
+  error: JsonRpcError = {
+    code: ErrorCodes.SIGNING_ADDRESS_NOT_FOUND,
+    message: 'Something went wrong'
+  };
+
+  constructor(public readonly id: number) {}
+
+  toJSON() {
+    return {
+      jsonrpc: this.jsonrpc,
+      id: this.id,
+      error: this.error
+    };
+  }
+}
+
+export class SigningAddressNotFoundError extends ChannelClientError {
+  error: JsonRpcError = {
+    code: ErrorCodes.SIGNING_ADDRESS_NOT_FOUND,
+    message: 'Signing address not found'
+  };
+}
+
+export class InvalidAppDefinitionError extends ChannelClientError {
+  error: JsonRpcError = {
+    code: ErrorCodes.INVALID_APP_DEFINITION,
+    message: 'Invalid app definition'
+  };
+}
+
+export class InvalidAppDataError extends ChannelClientError {
+  error: JsonRpcError = {
+    code: ErrorCodes.INVALID_APP_DATA,
+    message: 'Invalid app data'
+  };
+}
+
+export class UnsupportedTokenError extends ChannelClientError {
+  error: JsonRpcError = {
+    code: ErrorCodes.UNSUPPORTED_TOKEN,
+    message: 'Unsupported token'
+  };
+}
+
+export class ChannelNotFoundError extends ChannelClientError {
+  error: JsonRpcError = {
+    code: ErrorCodes.CHANNEL_NOT_FOUND,
+    message: 'Channel not found'
+  };
+}
+
+export const ErrorCodesToObjectsMap: {[key in ErrorCodes]: typeof ChannelClientError} = {
+  [ErrorCodes.CHANNEL_NOT_FOUND]: ChannelNotFoundError,
+  [ErrorCodes.INVALID_APP_DATA]: InvalidAppDataError,
+  [ErrorCodes.INVALID_APP_DEFINITION]: InvalidAppDefinitionError,
+  [ErrorCodes.SIGNING_ADDRESS_NOT_FOUND]: SigningAddressNotFoundError,
+  [ErrorCodes.UNSUPPORTED_TOKEN]: UnsupportedTokenError
+};
